+++ src/data/departmentData.js
@@ -1,0 +1,131 @@
+import cardiology from '../assets/images/departmentsphoto/cardiology.jpg';
+import nephrology from '../assets/images/departmentsphoto/nephrology.jpg';
+import Orthopedics from '../assets/images/departmentsphoto/orthopedic.jpg';
+import generalsurgery from '../assets/images/departmentsphoto/General-Surger.jpg';
+import generalmedicine from '../assets/images/departmentsphoto/General-Medicine.jpg';
+import neuroscience from '../assets/images/departmentsphoto/neuroscience.jpg';
+import gynaecology from '../assets/images/departmentsphoto/gyanocology.png';
+import paediatrics from '../assets/images/departmentsphoto/paedaitrics.jpg';
+import ENT from '../assets/images/departmentsphoto/ent.jpg';
+import Oncology from '../assets/images/departmentsphoto/oncology.jpg';
+import Dermatology from '../assets/images/departmentsphoto/dermatology.jpg';
+import Cardiology from '../assets/images/departmentsphoto/cardiology.jpg';
+import Nephrology from '../assets/images/departmentsphoto/nephrology.jpg';
+import Anesthesiology from '../assets/images/departmentsphoto/anesthesiology.png';
+
+const departmentData = [
+  {
+    id: 1,
+    name: 'Cardiology',
+    nepali: 'कार्डियोलोजी',
+    description:
+      'Expert care for heart conditions with advanced diagnostics and treatments.',
+    image: cardiology,
+  },
+  {
+    id: 2,
+    name: 'Neurology',
+    nepali: 'स्नायु विज्ञान',
+    description:
+      'Comprehensive care for neurological disorders with a team of specialists.',
+    image: nephrology,
+  },
+  {
+    id: 3,
+    name: 'Orthopedics',
+    nepali: 'हाड जोर्नी विभाग',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: Orthopedics,
+  },
+  {
+    id: 4,
+    name: 'General Surgery',
+    nepali: 'शल्य चिकित्सा',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: generalsurgery,
+  },
+  {
+    id: 5,
+    name: 'General Medicine',
+    nepali: 'सामान्य चिकित्सा',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: generalmedicine,
+  },
+  {
+    id: 6,
+    name: 'Neuroscience',
+    nepali: 'स्नायु विज्ञान',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: neuroscience,
+  },
+  {
+    id: 7,
+    name: 'Gynaecology/ Obstetrics',
+    nepali: 'स्त्री रोग / प्रसूति विज्ञान',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: gynaecology,
+  },
+  {
+    id: 8,
+    name: 'Paediatrics',
+    nepali: 'बालरोग',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: paediatrics,
+  },
+  {
+    id: 9,
+    name: 'ENT',
+    nepali: 'कान नाक घाँटी',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: ENT,
+  },
+  {
+    id: 10,
+    name: 'Cardiology',
+    nepali: 'कार्डियोलोजी',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: Cardiology,
+  },
+  {
+    id: 11,
+    name: 'Oncology',
+    nepali: 'ओन्कोलोजी',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: Oncology,
+  },
+  {
+    id: 12,
+    name: 'Dermatology',
+    nepali: 'त्वचा विज्ञान',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: Dermatology,
+  },
+  {
+    id: 13,
+    name: 'Nephrology',
+    nepali: 'नेफ्रोलोजी',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: Nephrology,
+  },
+  {
+    id: 14,
+    name: 'Anesthesiology',
+    nepali: 'एनेस्थेसियोलोजी',
+    description:
+      'Specialized treatment for musculoskeletal issues, including surgeries and rehabilitation.',
+    image: Anesthesiology,
+  },
+];
+
+export default departmentData;
